Student record management system
Rather than storing the results of more than 1000 students manually ,this system makes it easier to manage and organise data thus reducing data redundancy ,inconsistency and thus making it more accessible and efficient. Incorporating a way in which on giving the enrollment number and branch the course gets assigned to that particular student makes the process much more easier.